feat: rewrite in rust  and improve kftray server (#123)
* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* feat: rewrite in rust  and improve kftray server

* fix: tcp forward to prepare to add another workload

+++ src-tauri/utils/kubeforward/src/kubecontext.rs
@@ -174,9 +174,7 @@ pub async fn list_service_ports(
                                         if cp.name.as_deref() == Some(name) {
                                             service_port_infos.push(KubeServicePortInfo {
                                                 name: cp.name.clone(),
-                                                port: Some(IntOrString::Int(
-                                                    cp.container_port as i32,
-                                                )),
+                                                port: Some(IntOrString::Int(cp.container_port)),
                                             });
                                             break 'port_search;
                                         }
